Move GoRoot.tsx to Go/ui folder
Consistent with structure of other mechanics
